# Two sorbet-runtime call-validation optimizations