--- a/gems/sorbet-runtime/lib/types/profile.rb
+++ b/gems/sorbet-runtime/lib/types/profile.rb
@@ -1,6 +1,7 @@
 # typed: true
 # frozen_string_literal: true
 
+# Deprecated, kept only for partial backwards compatibility
 module T::Profile
   SAMPLE_RATE = 101 # 1 out of that many typechecks will be measured
   class <<self
@@ -8,23 +9,17 @@ module T::Profile
     attr_accessor :typecheck_samples
     attr_accessor :typecheck_sample_attempts
     def typecheck_duration_estimate
-      total_typechecks = typecheck_samples * SAMPLE_RATE + (SAMPLE_RATE - typecheck_sample_attempts)
-      typechecks_measured = typecheck_samples * SAMPLE_RATE
-      if typechecks_measured.positive?
-        typecheck_duration * SAMPLE_RATE * 1.0 * total_typechecks / typechecks_measured
-      else
-        0.0
-      end
+      0.0
     end
 
     def typecheck_count_estimate
-      typecheck_samples * SAMPLE_RATE
+      0
     end
 
     def reset
       @typecheck_duration = 0
       @typecheck_samples = 0
-      @typecheck_sample_attempts = SAMPLE_RATE
+      @typecheck_sample_attempts = 0
     end
   end
   self.reset