… gems/sorbet-runtime/.ruby-version
… gems/sorbet-runtime/lib/types/private/methods/call_validation_2_7.rb
… gems/sorbet-runtime/test/types/profile.rb
… gems/sorbet-runtime/tools/generate_call_validation.cc