Two changes:
1. Deprecate T::Profile, and rip out the sampled timing code from call validation.
2. Add a new semi-fast path for call validation with up to four positional arguments and no keyword arguments or other special cases, very similar to the existing fast path except without the restriction that types must be T::Types::Simple. (~~Also method-vs-procedure is left for a runtime check, for now, to avoid blowing up code size and because one extra branch didn't seem like a big deal relative to the rest.~~ edit: when merging the generator code, it was simpler to just leave this distinction in for both paths.)

These are really related only in intent; I actually implemented T::Profile support in the new fast-ish path before ripping it out. So I'm happy to split into separate PRs if folks prefer. 

### Motivation
1. I don't believe this timer measures anything useful. It doesn't account for a significant proportion of runtime typechecking overhead, probably the majority of overhead, because (a) it doesn't capture the impact of adding the `define_method` stack frame itself, and (b) it doesn't capture any impact of `T.let`, `T.must`, `T.unsafe` etc. A tool like StackProf will give a much better picture, and (if configured well) with much lower measurement overhead. Microbenchmarks show >20% speedup on single-argument methods.

2. I don't have hard statistics but my impression is that a large number of relatively-simple methods fail to qualify for the existing fast path because they use something like T::Boolean, T.nilable, or T.untyped, even as just one argument among otherwise simple ones. Microbenchmarks show a >40% speedup on single-nilable-argument methods. 

`bundle exec rake bench:typecheck` results for 2.7, slightly abridged, on:

master:
```
T::Types::Simple#valid?: 41.252 ns
T::Types::Union#valid?: 139.641 ns
T.let(..., Integer): 428.153 ns
sig {params(x: Integer).void}: 337.683 ns
T.let(..., T.nilable(Integer)): 1.036 μs
sig {params(x: T.nilable(Integer)).void}: 862.766 ns
T.let(..., Example): 409.354 ns
sig {params(x: Example).void}: 307.338 ns
T.let(..., T.nilable(Example)): 1.043 μs
sig {params(x: T.nilable(Example)).void}: 875.392 ns
sig {params(s: Symbol, x: Integer, y: Integer).void} (with kwargs): 1.504 μs
```

just the new fast path:
```
T::Types::Simple#valid?: 42.637 ns
T::Types::Union#valid?: 142.595 ns
T.let(..., Integer): 389.78 ns
sig {params(x: Integer).void}: 330.284 ns
T.let(..., T.nilable(Integer)): 1.004 μs
sig {params(x: T.nilable(Integer)).void}: 492.045 ns
T.let(..., Example): 396.391 ns
sig {params(x: Example).void}: 313.137 ns
T.let(..., T.nilable(Example)): 1.01 μs
sig {params(x: T.nilable(Example)).void}: 500.454 ns
sig {params(s: Symbol, x: Integer, y: Integer).void} (with kwargs): 1.507 μs
```

with both changes:
```
T::Types::Simple#valid?: 40.547 ns
T::Types::Union#valid?: 161.091 ns
T.let(..., Integer): 405.242 ns
sig {params(x: Integer).void}: 215.672 ns
T.let(..., T.nilable(Integer)): 1.038 μs
sig {params(x: T.nilable(Integer)).void}: 377.882 ns
T.let(..., Example): 393.038 ns
sig {params(x: Example).void}: 207.139 ns
T.let(..., T.nilable(Example)): 1.078 μs
sig {params(x: T.nilable(Example)).void}: 394.03 ns
sig {params(s: Symbol, x: Integer, y: Integer).void} (with kwargs): 1.424 μs
```

(Only the changes in the `sig` results could be meaningful, but I've left the others in as points of comparison & indicators of the noise level.)

### Test plan
Relying on existing tests