@jez 2021-04-18
Do the api method tests use T::Profile right now in Stripe codebase? Is the plan to delete or change those tests?