@jez 2021-04-19
> T::Profile is only used in a couple of production log-lines and I don't believe it provides real value there

Gotcha, so it seems like it's basically just populating some log lines, and since you've made a case that those log lines are not even really accurate, that it's not worth keeping these things around.